[v0.5.0] Example replays.
They can be attached to an uploaded scheme (it would be silly to attach
a replay to a scheme we're creating, so I don't plan to add this feature
unless there's a massive enough request - it's easy anyway, it's all
about copy-pasting). They are also downloadable, from the schemes list.

Next step, being able to attach them to an already existing scheme
(either the creator has to add them on his own, and the scheme password
would be required; or anyone can upload replays and there's an
approve/disapprove interface for the creator, the scheme password being
required there again).