[v0.7.5] An improvement and a few bugfixes.
From the changelog:

- [Improvement] It is now possible to rename a scheme you're Uploading.
- [Fixed bug] On the Scheme Uploading page, a <td> tag was closed, but
never opened. It has been removed. (Note that this doesn't affect the
displayed page; it only made the page invalid according to the W3C XHTML
1.0 Strict standards.)
- [Fixed bug] Due to an error, the paging system on the Scheme List
didn't work: it would only show the page with the oldest schemes. Note
that this didn't affect the public version of the scheme editor, because
at that time schemes were still fitting on a single page; the bug has
been noticed on my local version.
- [Fixed bug] Due to another error, the paging system on the Scheme List
would be affected by example replays if more than one was attached to
any of the displayed schemes: for example, only 27 schemes (instead of
30) would be shown on a page where 4 example replays were attached to
the same scheme.