[v1.2.3-4] Some bugfixes.
1.2.3 - 02.06.2013:
- [Fixed bug] The ability to rename a scheme while editing it,
introduced in v1.2.2, broke scheme editing, in that the scheme generated
lacked the author's name in the file name; as a result, the Scheme
Viewer would read the originally created/uploaded scheme. (This was due
to the query overwriting data from the first one.) Credits go to
Star&Moon for reporting the bug.
- [Fixed] Fields in the Scheme Uploading form were inconsistently padded
(i.e. some were padded and some weren't). The padding has been removed.

1.2.4 - 11.06.2013:
- [Fixed bug] The following v1.1.0 feature: "It is now possible to
create a scheme without saving it on the server and in the database."
was overdone and thus didn't work properly (the generated file was
incorrect). Now, you can successfully create a scheme without having it
being published.
- [Fixed bug] Viewing a v1 scheme (221 bytes long scheme files) will no
longer output errors.